Fixed the plugin not loading bug, here is what happened: - The logger was accidentally delted from the constructor - There were a bunch of typos in bsml and the ViewController that made the ViewController fail to load - I also renamed the ViewController and View to make the naming scheme more consistent - I changed the modifier tag to the bool setting tag because it looked better that way imo
