+++ NoBoom/NoBoomInstaller.cs
@@ -1,18 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Zenject;
+﻿using Zenject;
 
-namespace NoBoom
+namespace NoBoom.Installers
 {
     class NoBoomInstaller : Installer
     {
-
         public override void InstallBindings()
         {
-            Container.Bind<NoBoomViewController>().AsSingle();
+            Container.BindInterfacesTo<ModifierViewController>().AsSingle();
         }
     }
 }
