+++ NoBoom/Views/ModifierView.bsml
@@ -1,0 +1,5 @@
+﻿<bg xmlns:xsi='http://www.w3.org/2001/XMLSchema-instance' xsi:schemaLocation='https://monkeymanboy.github.io/BSML-Docs/ https://raw.githubusercontent.com/monkeymanboy/BSML-Docs/gh-pages/BSMLSchema.xsd'>
+  <modifier-container child-expand-width='true' horizontal-fit='PreferredSize'>
+    <bool-setting value='enabled' text='Enabled' />
+  </modifier-container>
+</bg>
